# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuantSynthica Market API",
    "description": "Unified Financial-Data API integrating yfinance, TradingView-Screener, and Screener.in with quantitative analytics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/fundamentals/{symbol}": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get unified company fundamentals",
        "operationId": "get_company_fundamentals_api_v1_fundamentals__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedFundamentalsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/income-statement": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get income statement / P&L",
        "operationId": "get_income_statement_api_v1_fundamentals__symbol__income_statement_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialStatementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/balance-sheet": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get balance sheet",
        "operationId": "get_balance_sheet_api_v1_fundamentals__symbol__balance_sheet_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialStatementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/cash-flow": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get cash flow statement",
        "operationId": "get_cash_flow_api_v1_fundamentals__symbol__cash_flow_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialStatementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/ratios": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get valuation and financial ratios",
        "operationId": "get_ratios_api_v1_fundamentals__symbol__ratios_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatiosResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/shareholding": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get shareholding pattern",
        "operationId": "get_shareholding_api_v1_fundamentals__symbol__shareholding_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareholdingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/institutional-holders": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get top institutional shareholders",
        "operationId": "get_institutional_holders_api_v1_fundamentals__symbol__institutional_holders_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/mutual-funds": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get mutual fund shareholders",
        "operationId": "get_mutual_fund_holders_api_v1_fundamentals__symbol__mutual_funds_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/insider-transactions": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get insider buy/sell transactions",
        "operationId": "get_insider_transactions_api_v1_fundamentals__symbol__insider_transactions_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsiderTransactionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/sustainability": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get ESG and sustainability risk ratings",
        "operationId": "get_sustainability_ratings_api_v1_fundamentals__symbol__sustainability_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SustainabilityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/peers": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get industry peer comparison table (from Screener.in)",
        "operationId": "get_peers_comparison_api_v1_fundamentals__symbol__peers_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/analysis": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get pros and cons analysis (from Screener.in)",
        "operationId": "get_corporate_analysis_api_v1_fundamentals__symbol__analysis_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorporateAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fundamentals/{symbol}/documents": {
      "get": {
        "tags": [
          "Fundamentals"
        ],
        "summary": "Get concalls, presentations, annual reports (from Screener.in)",
        "operationId": "get_corporate_documents_api_v1_fundamentals__symbol__documents_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorporateAnalysisResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "pros": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pros"
          },
          "cons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cons"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "metadata"
        ],
        "title": "CorporateAnalysisResponse"
      },
...
      "DocumentsResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentItem"
            },
            "type": "array",
            "title": "Documents"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "count",
          "metadata"
        ],
        "title": "DocumentsResponse"
      },
...
      "FinancialStatementData": {
        "properties": {
          "statement_type": {
            "type": "string",
            "title": "Statement Type"
          },
          "periods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Periods"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/FinancialRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "raw_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Data"
          }
        },
        "type": "object",
        "required": [
          "statement_type"
        ],
        "title": "FinancialStatementData"
      },
      "FinancialStatementResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "statement_type": {
            "type": "string",
            "title": "Statement Type"
          },
          "data": {
            "$ref": "#/components/schemas/FinancialStatementData"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "statement_type",
          "data",
          "metadata"
        ],
        "title": "FinancialStatementResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HoldersResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "holders": {
            "items": {
              "$ref": "#/components/schemas/HolderItem"
            },
            "type": "array",
            "title": "Holders"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "type",
          "metadata"
        ],
        "title": "HoldersResponse"
      },
...
      "InsiderTransactionsResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/InsiderTransactionItem"
            },
            "type": "array",
            "title": "Transactions"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "count",
          "metadata"
        ],
        "title": "InsiderTransactionsResponse"
      },
...
      "MultiSourceMetadata": {
        "properties": {
          "sources": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Sources",
            "description": "Field to provider mapping"
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "successful_providers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Successful Providers"
          },
          "failed_providers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Failed Providers"
          }
        },
        "type": "object",
        "title": "MultiSourceMetadata"
      },
...
      "PeersResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "peers": {
            "items": {
              "$ref": "#/components/schemas/PeerItem"
            },
            "type": "array",
            "title": "Peers"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "metadata"
        ],
        "title": "PeersResponse"
      },
...
      "RatiosResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "data": {
            "$ref": "#/components/schemas/RatiosData"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "data",
          "metadata"
        ],
        "title": "RatiosResponse"
      },
...
      "ShareholdingItem": {
        "properties": {
          "holder_category": {
            "type": "string",
            "title": "Holder Category"
          },
          "values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "holder_category"
        ],
        "title": "ShareholdingItem"
      },
      "ShareholdingResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "periods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Periods"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/ShareholdingItem"
            },
            "type": "array",
            "title": "Categories"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "metadata"
        ],
        "title": "ShareholdingResponse"
      },
...
      "SustainabilityResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ESGRatingsData"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "metadata"
        ],
        "title": "SustainabilityResponse"
      },
...
      "UnifiedFundamentalsData": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "about": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "About"
          },
          "ratios": {
            "$ref": "#/components/schemas/RatiosData"
          },
          "quarterly_results": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatementData"
              },
              {
                "type": "null"
              }
            ]
          },
          "profit_and_loss": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatementData"
              },
              {
                "type": "null"
              }
            ]
          },
          "balance_sheet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatementData"
              },
              {
                "type": "null"
              }
            ]
          },
          "cash_flow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatementData"
              },
              {
                "type": "null"
              }
            ]
          },
          "shareholding": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ShareholdingItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shareholding"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "ratios"
        ],
        "title": "UnifiedFundamentalsData"
      },
      "UnifiedFundamentalsResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "data": {
            "$ref": "#/components/schemas/UnifiedFundamentalsData"
          },
          "metadata": {
            "$ref": "#/components/schemas/MultiSourceMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "data",
          "metadata"
        ],
        "title": "UnifiedFundamentalsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}